requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pandas==2.2.2
numpy==1.26.4
openai==1.51.0
pydantic==2.9.2
python-dotenv==1.0.1
plotly==5.24.1
scipy==1.14.1
scikit-learn==1.5.2
